refactor scripts and add organism dispply.
move scripts to the script.js folder and refactor the mineNode method to use tables rather than simple a tags, so we can display organisms. At this point also changed to a less verbose way of generating the content for the mine names. Given that these pages are very simple and not likely to need much modification over time I've gone for the simple string concatenation innerHTML approach - not something I'd recommend for complex applications but totally suitable to our simple needs here.